Initial upload of libhio.
This commit is the first commit for libhio. With this commit libhio
is at v0.1 alpha0. It supports n->n->n and n->1->n IO where the
data read on read is the same data written by the same process on
write. As development continues support will be added for full
n->1->m IO.