Add support for dynamic modules in <prefix>/lib/hio_version/modules
This commit adds support for discovering/loading dynamic hio modules
from the modules directory in the hio prefix. For example, for hio 1.0
modules can be placed in:

<prefix>/lib/hio_1.0/modules

and they will be scanned and loaded. At this time no external modules
currently exist.

This commit also cleans up the posix code to remove the use of the
optional PATH_MAX field. Instead, use asprintf to allocate memory for
paths.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>