Fix bug in posix element_open
This commit fixes an outstanding bug in element open for datasets
opened in basic mode. The code used the "w" file mode for fopen which
would cause the file to be truncated if it already existed. This was
not the desired behavior. To fix this bug a combination of open and
fdopen is used to open the file for writing, creating it if needed,
without truncating it.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>