Update version to 1.0
With this initial release most of the API is complete with the
exception of the restart recommendation feature. That feature will
come in the 1.1 release.

Basic I/O mode is fully supported as of the 1.0 release. Optimized I/O
will come in a future release and will require a thread-safe MPI
implementation. Optimized mode will add support for refactored and
turnstiled I/O. This will be targeted for Trinity but will also
support TLCC1 and TLCC2 hardware.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>