Add support for optimized mode on shared files
This commit adds support for chunking data in shared elements into
multiple data files. The data file used for a specific offset is based
on the block size for the dataset. By default the block size used is
based on the number of ranks performing I/O and an estimated 2 GiB of
memory per rank.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>