config: update semantics of hio_config_set/get_value()
This commit updates the semantics of the configuration function when
setting/reading an unknown configuration variable. If a variable does
not exist when hio_config_set_value() is called on an object the key
and value will be stored in the object. Subsequent calls to
hio_config_get_value() before the value is used will return the set
value.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>