posix: set default stripe count
This commit fixes a bug that occurs when the dataset directory is
using the default striping count. In this case the ioctl may return
lmm_stripe_count = 0. Since hio assumes a stripe count >= 1 this was
triggering a problem when setting up the shared memory segment. To be
safe always set the default to 1 (or larger if writing shared files).

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>