posix: improve support for optimized file mode
This commit further improved the optimized file mode provided by the
posix backend. This mode causes all element data written on a node to
be written to the same backing file. This requires building a map of
the data. Before this commit the map was loaded onto each process so
all the data could be accessed by any rank. The problem with this
approach is the map could grow very large causing the process to run
out of memory. After this commit each IO leader (lowest rank on each
node) loads a portion of the map and uses the data to generate a
distributed hash table from the map using an MPI memory window. If a
process can not find the data in it's local map it can look into the
distributed hash table.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>