hio/manifest: store configuration in manifest
This commit adds the dataset configuration into the manifest. This
includes the stripe size, stripe_width, block size, etc. Some
configuration values already existed in the manifest. Those have been
removed in favor of the configuration values. The new manifest version
is 3.0 but the release is compatible with manifest v2.0.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>