posix: fix race in setting file striping
This commit fixes a race between a process creating a data file and
other processes opening the data file for writing. Before this commit
the striping was set *after* the file was created. This may be after
another process has started writing to the file. This will cause the
default striping to be used on some files. To prevent the data race
all dataset data is now stored in a subdirectory called "data". The
striping on this directory is set on dataset open.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>