datawarp: set staging properties of output directory
This commit attempts to fix a weirdness when using DataWarp. DataWarp
does not set the striping properties on any staged files. This will
leave us with the default striping for some potentially large output
files. To rectify this modify the lustre striping of the data
directory before starting the stage. This should catch all of the
large files in the typical case.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>