datawarp: keep track of dataset locations
This commit fixes an issue that can occur when canceling a stage on a
dataset where the dataset_uri was specified by the user. If the user
specified a URI with a '/' then the dataset unlink would always
fail. This commit fixes the error by storing the dataset header and
using the posix backend's unlink_dir to blow away the dataset.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>