datawarp: fix support for persistent datawarp mounts
This commit fixes an error that would cause the builtin datawarp
support from working when a persistent mount was detected. The
underlying cause was due to expecting the DW_JOB_STRIPED variable to
always be defined. This isn't the case unless there is a jobdw
directive in the run script.

Fixes #37

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>