spack: fix for spack to work on non-cray systems
or when one wants to use Open MPI on a cray system with spack.
The problem is that when using an MPI with spack, its going to
use a long value for the CC and CXX variables.  This confused
the libhio configury, resulting in it being built without mpi
support, leading to failure of xexec to properly compile, etc.

This commit fixes that problem.

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>