core: add a new module dataset unlink helper function
This commit adds a new helper function:
hio_module_dataset_unlink_dataset_header_fn_t

This function takes a description of the dataset that will be
removed. This description includes the name, set id, and path. The new
function is now used by default when removing an existing dataset when
HIO_FLAG_TRUNC is specified. This ensures that if the user specified a
path for the dataset we clean up the correct directory.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>